Deconvolve ST data set with matched scRNAseq data

Estimate the fraction of cell lineage and sub lineage.

Usage

SpaCET.deconvolution.matched.scRNAseq(
  SpaCET_obj,
  sc_includeMalignant = TRUE,
  cancerType,
  sc_counts,
  sc_annotation,
  sc_lineageTree,
  sc_downsampling = TRUE,
  sc_nCellEachLineage = 100,
  coreNo = 6
)

Arguments

SpaCET_obj

A SpaCET object.

sc_includeMalignant

Indicate whether the single cell data include malignant cells. If FALSE, please input a cancer type and then SpaCET will infer the malignant cell fraction based on its build-in reference.

cancerType

Cancer type of the current tumor ST sample.

sc_counts

Single cell count matrix with gene name (row) x cell ID (column).

sc_annotation

Single cell annotation matrix. This matrix should include two columns, i,e., cellID and cellType. Each row represents a single cell.

sc_lineageTree

Cell lineage tree. This should be organized by using a list, and the name of each element are major lineages while the value of elements are the corresponding sublineages. If a major lineage does not have any sublineages, the value of this major lineage should be itself.

sc_downsampling

Indicate whether downsample the single cell data for each cell type. Default: TRUE.

sc_nCellEachLineage

Cell count each lineage. Default: 100. If a cell type is comprised of >100 cells, only 100 cells per cell identity are randomly selected to generate cell type reference.

coreNo

Core number.

Value

A SpaCET object

Examples

SpaCET_obj <- SpaCET.deconvolution.matched.scRNAseq(SpaCET_obj, sc_counts, sc_annotation, sc_lineageTree)